refactor: extracted api calls o separate functions
