+++ src/pages/Results/Results.tsx
@@ -1,6 +1,7 @@
 import axios from "axios";
 import { useEffect } from "react";
 import { Link } from "react-router-dom";
+import { SaveToLeaderboard } from "../../ApiCalls/leaderboard";
 import { BASE_URL } from "../../ApiUrls/ApiUrls";
 import { useQuizPerformance } from "../../context/quizPerformance.context";
 import { useResults } from "../../context/quizResults.context";
@@ -11,20 +12,21 @@ export function Results() {
   console.log("quizPerformance ", quizPerformance);
   useEffect(() => {
     async function Run() {
-      const userPerformanceData = {
-        userData: {
-          name: localStorage.getItem("username"),
-          quizName: quizPerformance.quizName,
-          score: quizPerformance.score,
-        },
-      };
-      const response = await axios.post(
-        BASE_URL + "/leaderboard",
-        userPerformanceData
-      );
-      if (response.data.status) {
-        console.log("user saved to leaderboard");
-      }
+      // const userPerformanceData = {
+      //   userData: {
+      //     name: localStorage.getItem("username"),
+      //     quizName: quizPerformance.quizName,
+      //     score: quizPerformance.score,
+      //   },
+      // };
+      // const response = await axios.post(
+      //   BASE_URL + "/leaderboard",
+      //   userPerformanceData
+      // );
+      // if (response.data.status) {
+      //   console.log("user saved to leaderboard");
+      // }
+      await SaveToLeaderboard(quizPerformance);
     }
     Run();
   }, []);
